add RBAC in require file service
