+++ elevate/views.py
@@ -14,6 +14,7 @@ from rest_framework import generics, status, permissions
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework.decorators import api_view
+from rest_framework.permissions import AllowAny
 from knox.models import AuthToken
 from knox.views import LogoutView as KnoxLogoutView
 from rest_framework.parsers import MultiPartParser, FormParser, JSONParser
@@ -36,6 +37,7 @@ import time
 import traceback
 from datetime import datetime
 from django.db import transaction
+from django.utils.timezone import now
 
 logger = logging.getLogger(__name__)
 
@@ -82,7 +84,8 @@ class LoginAPI(generics.GenericAPIView):
                 user.profile.save()
             response_data = {
                 "user": UserSerializer(user, context=self.get_serializer_context()).data,
-                "token": AuthToken.objects.create(user)[1]
+                "token": AuthToken.objects.create(user)[1],
+                "message": "Login successful"
             }
             logger.info(f"Login successful for {user.username}")
             return Response(response_data, status=status.HTTP_200_OK)
@@ -127,10 +130,10 @@ class InviteUserAPI(APIView):
         try:
             with transaction.atomic():
                 user = User.objects.create_user(
-                    username=username, email=email, password=random_password, is_staff=True)
+                    username=username, email=email, password=random_password, is_staff=False)
                 UserProfile.objects.create(
                     user=user, organization=organization, invitation_status='Pending',
-                    invitation_username=username, temporary_password=random_password, invitation_token=token)
+                    invitation_username=username, temporary_password=random_password, invitation_token=token, role='user')
                 activation_link = f"{settings.FRONTEND_URL}/signin/{token}"
                 context = {
                     'username': username,
@@ -185,68 +188,115 @@ class SetPasswordAPI(generics.GenericAPIView):
 
 
 class PasswordResetRequestAPI(APIView):
-    permission_classes = [permissions.IsAdminUser]  # Restricted to admins
+    permission_classes = [permissions.AllowAny]
 
     def post(self, request):
         serializer = PasswordResetSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
-        email = serializer.validated_data['email']
+        email = serializer.validated_data['email'].lower().strip()
+
+        ALLOWED_PASSWORD_RESET_ROLES = ['admin', 'superuser']
+
         try:
             user = User.objects.get(email=email)
-            if not user.is_staff:
-                return Response({"error": "Password reset is only allowed for admin users"}, status=status.HTTP_403_FORBIDDEN)
+
+            if user.is_superuser and not hasattr(user, 'profile'):
+                pass 
+
+            else:
+                if not hasattr(user, 'profile'):
+                    return Response({"error": "User profile is missing"}, status=status.HTTP_400_BAD_REQUEST)
+
+                profile = user.profile
+
+                if profile.invitation_status.lower() == 'pending':
+                    return Response(
+                        {"error": "Invited users cannot request password reset"},
+                        status=status.HTTP_403_FORBIDDEN
+                    )
+
+                if profile.role.lower() not in ALLOWED_PASSWORD_RESET_ROLES:
+                    return Response(
+                        {"error": "Only admin users can request password reset"},
+                        status=status.HTTP_403_FORBIDDEN
+                    )
+
             token = get_random_string(length=32)
-            user.profile.password_reset_token = token  # Use new field
-            user.profile.save()
+            if hasattr(user, 'profile'):
+                user.profile.password_reset_token = token
+                user.profile.save()
+
             frontend_url = getattr(
                 settings, 'FRONTEND_URL', 'http://localhost:5173')
-            reset_link = f"{frontend_url}/password-reset/{urlsafe_base64_encode(force_bytes(user.pk))}/{token}/"
-            context = {
-                'reset_link': reset_link,
-                'current_year': datetime.now().year,
-            }
+            uid = urlsafe_base64_encode(force_bytes(user.pk))
+            reset_link = f"{frontend_url}/password-reset/{uid}/{token}/"
+
+            context = {'reset_link': reset_link, 'current_year': now().year}
             html_message = render_to_string(
                 'emails/password_reset.html', context)
-            email = EmailMessage(
+
+            email_message = EmailMessage(
                 subject='Elevate Password Reset',
                 body=html_message,
                 from_email=settings.DEFAULT_FROM_EMAIL,
                 to=[email],
             )
-            email.content_subtype = "html"
-            email.send()
-            return Response({"message": "Password reset email sent"}, status=status.HTTP_200_OK)
-        except User.DoesNotExist:
-            return Response({"error": "User with this email does not exist"}, status=status.HTTP_404_NOT_FOUND)
-        except Exception as e:
-            logger.error(f"Password reset email failed: {str(e)}")
-            return Response({"error": "Failed to send password reset email"}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+            email_message.content_subtype = "html"
+            email_message.send()
 
+            return Response({"message": "Password reset email sent"}, status=status.HTTP_200_OK)
+
+        except User.DoesNotExist:
+            return Response({"error": "User not found"}, status=status.HTTP_404_NOT_FOUND)
+
+        except Exception:
+            return Response({"error": "Failed to send password reset email"}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+        
 
 class PasswordResetAPI(generics.GenericAPIView):
     serializer_class = PasswordResetNoEmailSerializer
-    permission_classes = [permissions.IsAdminUser]  # Restricted to admins
+    # ✅ Allow unauthenticated access
+    permission_classes = [permissions.AllowAny]
 
     def post(self, request, *args, **kwargs):
         serializer = self.get_serializer(data=request.data)
         serializer.is_valid(raise_exception=True)
+
         uidb64 = serializer.validated_data['uidb64']
         token = serializer.validated_data['token']
         new_password = serializer.validated_data['new_password']
+
         try:
             uid = force_str(urlsafe_base64_decode(uidb64))
             user = User.objects.get(pk=uid)
+
             if not user.is_staff:
-                return Response({"error": "Password reset is only allowed for admin users"}, status=status.HTTP_403_FORBIDDEN)
-            if user.profile.password_reset_token == token:  # Use new field
+                return Response(
+                    {"error": "Password reset is only allowed for admin users"},
+                    status=status.HTTP_403_FORBIDDEN
+                )
+
+            # ✅ Case 1: Superuser without profile
+            if user.is_superuser and not hasattr(user, "profile"):
+                user.set_password(new_password)
+                user.save()
+                return Response({"message": "Password reset successfully"}, status=status.HTTP_200_OK)
+
+            # ✅ Case 2: Admin user with profile and matching token
+            if hasattr(user, "profile") and user.profile.password_reset_token == token:
                 user.set_password(new_password)
                 user.profile.password_reset_token = ''
                 user.profile.save()
                 user.save()
                 return Response({"message": "Password reset successfully"}, status=status.HTTP_200_OK)
+
             return Response({"error": "Invalid token"}, status=status.HTTP_400_BAD_REQUEST)
+
         except (User.DoesNotExist, ValueError):
             return Response({"error": "Invalid user or token"}, status=status.HTTP_400_BAD_REQUEST)
+
+        except Exception as e:
+            return Response({"error": f"Unexpected error: {str(e)}"}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
 
 def password_reset_confirm(request, uidb64, token):
@@ -255,7 +305,7 @@ def password_reset_confirm(request, uidb64, token):
         user = User.objects.get(pk=uid)
         if not user.is_staff:
             return redirect(f"{getattr(settings, 'FRONTEND_URL', 'http://localhost:5173')}/password-reset/invalid/")
-        if user.profile.password_reset_token == token:  # Use new field
+        if user.profile.password_reset_token == token:
             return redirect(f"{getattr(settings, 'FRONTEND_URL', 'http://localhost:5173')}/password-reset/{uidb64}/{token}/")
         return redirect(f"{getattr(settings, 'FRONTEND_URL', 'http://localhost:5173')}/password-reset/invalid/")
     except (User.DoesNotExist, ValueError):
@@ -263,22 +313,51 @@ def password_reset_confirm(request, uidb64, token):
 
 
 class PasswordResetLinkValidateAPI(APIView):
-    permission_classes = [permissions.IsAdminUser]  # Restricted to admins
+    permission_classes = [permissions.AllowAny]
 
     def get(self, request, uidb64, token):
         try:
             uid = force_str(urlsafe_base64_decode(uidb64))
             user = User.objects.get(pk=uid)
-            if not user.is_staff:
-                return Response({"error": "Password reset is only allowed for admin users"}, status=status.HTTP_403_FORBIDDEN)
-            if user.profile.password_reset_token == token:  # Use new field
-                return Response({"message": "Valid reset link"}, status=status.HTTP_200_OK)
-            return Response({"error": "Invalid reset link"}, status=status.HTTP_400_BAD_REQUEST)
-        except (User.DoesNotExist, ValueError):
-            return Response({"error": "Invalid reset link"}, status=status.HTTP_400_BAD_REQUEST)
+
+            # ✅ Case 1: Superuser without profile
+            if user.is_superuser and not hasattr(user, 'profile'):
+                return Response({"valid": True}, status=status.HTTP_200_OK)
+
+            # ✅ Case 2: User with profile and matching token
+            if hasattr(user, "profile"):
+                stored_token = user.profile.password_reset_token
+                if stored_token == token:
+                    return Response({"valid": True}, status=status.HTTP_200_OK)
+                else:
+                    return Response({
+                        "valid": False,
+                        "error": "Invalid or expired token."
+                    }, status=status.HTTP_400_BAD_REQUEST)
+
+            # ❌ Case 3: No profile, and not a superuser
+            return Response({
+                "valid": False,
+                "error": "User is not allowed to reset password."
+            }, status=status.HTTP_403_FORBIDDEN)
+
+        except User.DoesNotExist:
+            return Response({
+                "valid": False,
+                "error": "User does not exist."
+            }, status=status.HTTP_400_BAD_REQUEST)
+
+        except Exception as e:
+            return Response({
+                "valid": False,
+                "error": f"Unexpected error: {str(e)}"
+            }, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+
 
 
 class GetInvitationDetailsAPI(APIView):
+    permission_classes = [permissions.AllowAny]
+
     def get(self, request, token):
         try:
             profile = UserProfile.objects.get(invitation_token=token)
@@ -308,8 +387,8 @@ class DeleteInvitedUserAPI(generics.DestroyAPIView):
     def perform_destroy(self, instance):
         instance.profile.delete()
         instance.delete()
- 
-        
+
+
 class AnalysisView(APIView):
     permission_classes = [permissions.IsAuthenticated]
     parser_classes = (MultiPartParser, FormParser, JSONParser)
@@ -322,14 +401,16 @@ class AnalysisView(APIView):
         file_path = None
         start_time = time.time()
         try:
-            logger.info(f"Starting EV analysis for user {request.user.username} at {start_time}")
+            logger.info(
+                f"Starting EV analysis for user {request.user.username} at {start_time}")
             all_data = request.data.copy()
             files = request.FILES
 
-            # Handle uploaded file (isLoadSplitFile)
             if 'isLoadSplitFile' in all_data and isinstance(all_data['isLoadSplitFile'], str):
-                file_relative_path = all_data['isLoadSplitFile'].replace('/media/', '')
-                file_path = os.path.join(settings.MEDIA_ROOT, file_relative_path)
+                file_relative_path = all_data['isLoadSplitFile'].replace(
+                    '/media/', '')
+                file_path = os.path.join(
+                    settings.MEDIA_ROOT, file_relative_path)
                 if not os.path.exists(file_path):
                     raise ValueError(f"File not found at path: {file_path}")
                 all_data['isLoadSplitFile'] = file_path
@@ -341,29 +422,30 @@ class AnalysisView(APIView):
                     raise ValueError("Only Excel files are supported")
                 fs = FileSystemStorage()
                 upload_folder = 'FileUpload'
-                os.makedirs(os.path.join(settings.MEDIA_ROOT, upload_folder), exist_ok=True)
-                filename = fs.save(os.path.join(upload_folder, file.name), file)
+                os.makedirs(os.path.join(settings.MEDIA_ROOT,
+                            upload_folder), exist_ok=True)
+                filename = fs.save(os.path.join(
+                    upload_folder, file.name), file)
                 file_path = fs.path(filename)
                 all_data['isLoadSplitFile'] = file_path
 
-            # Set default name for invited users if not provided
             if 'name' not in all_data:
                 all_data['name'] = request.user.username
-                logger.info(f"Set default name to {request.user.username} for invited user")
+                logger.info(
+                    f"Set default name to {request.user.username} for invited user")
 
-            # Process categories
-            processed_data, created_load_categories, created_vehicle_categories = self.process_categories(all_data)
+            processed_data, created_load_categories, created_vehicle_categories = self.process_categories(
+                all_data)
             all_data.update(processed_data)
             all_data['user_name'] = request.user.username
 
-            # Serialize and save instance
-            serializer = AnalysisSerializer(data=all_data, context={'request': request})
+            serializer = AnalysisSerializer(
+                data=all_data, context={'request': request})
             if not serializer.is_valid():
                 logger.error(f"Serializer errors: {serializer.errors}")
                 return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
             ev_instance = serializer.save(user=request.user)
 
-            # Log analysis start
             user_analysis_log = UserAnalysis.objects.create(
                 userName=request.user.username,
                 status='Processing',
@@ -371,11 +453,10 @@ class AnalysisView(APIView):
                 time=0.0
             )
 
-            # Prepare and run analysis
             analysis_data = self.prepare_analysis_data(ev_instance)
-            results = self.run_full_analysis(analysis_data, str(ev_instance.id))
+            results = self.run_full_analysis(
+                analysis_data, str(ev_instance.id))
 
-            # Update log and cleanup
             user_analysis_log.status = 'Completed'
             user_analysis_log.time = time.time() - start_time
             user_analysis_log.save()
@@ -383,23 +464,27 @@ class AnalysisView(APIView):
             return Response(results, status=status.HTTP_201_CREATED)
 
         except Exception as e:
-            logger.error(f"Analysis failed: {str(e)}\n{traceback.format_exc()}")
+            logger.error(
+                f"Analysis failed: {str(e)}\n{traceback.format_exc()}")
             UserAnalysis.objects.create(
                 userName=request.user.username,
                 status='Failed',
                 errorLog=f"{str(e)}\n{traceback.format_exc()}",
                 time=time.time() - start_time
             )
-            self.cleanup_temporary_data(created_load_categories, created_vehicle_categories, ev_instance, file_path)
+            self.cleanup_temporary_data(
+                created_load_categories, created_vehicle_categories, ev_instance, file_path)
             return Response({'error': str(e), 'traceback': traceback.format_exc()}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
         finally:
-            self.cleanup_temporary_data(created_load_categories, created_vehicle_categories, None, file_path)
+            self.cleanup_temporary_data(
+                created_load_categories, created_vehicle_categories, None, file_path)
 
     def process_categories(self, all_data):
         processed_data = {}
         created_load_categories = []
         created_vehicle_categories = []
-        load_categories = all_data.get('category_data', []) or all_data.get('categoryData', [])
+        load_categories = all_data.get(
+            'category_data', []) or all_data.get('categoryData', [])
         if len(load_categories) > 6:
             raise ValueError("Maximum 6 load categories allowed")
         for idx, category_data in enumerate(load_categories[:6], start=1):
@@ -407,12 +492,15 @@ class AnalysisView(APIView):
             try:
                 serializer.is_valid(raise_exception=True)
             except Exception as e:
-                logger.error(f"Load category {idx} ({category_data.get('category', 'unknown')}) validation failed: {str(e)}")
-                raise ValueError(f"Load category {idx} ({category_data.get('category', 'unknown')}) validation failed: {str(e)}")
+                logger.error(
+                    f"Load category {idx} ({category_data.get('category', 'unknown')}) validation failed: {str(e)}")
+                raise ValueError(
+                    f"Load category {idx} ({category_data.get('category', 'unknown')}) validation failed: {str(e)}")
             category = serializer.save()
             created_load_categories.append(category)
             processed_data[f'loadCategory{idx}'] = category.id
-        vehicle_categories = all_data.get('vehicle_category_data', []) or all_data.get('vehicleCategoryData', [])
+        vehicle_categories = all_data.get(
+            'vehicle_category_data', []) or all_data.get('vehicleCategoryData', [])
         if len(vehicle_categories) > 5:
             raise ValueError("Maximum 5 vehicle categories allowed")
         for idx, vehicle_data in enumerate(vehicle_categories[:5], start=1):
@@ -420,8 +508,10 @@ class AnalysisView(APIView):
             try:
                 serializer.is_valid(raise_exception=True)
             except Exception as e:
-                logger.error(f"Vehicle category {idx} ({vehicle_data.get('vehicleCategory', 'unknown')}) validation failed: {str(e)}")
-                raise ValueError(f"Vehicle category {idx} ({vehicle_data.get('vehicleCategory', 'unknown')}) validation failed: {str(e)}")
+                logger.error(
+                    f"Vehicle category {idx} ({vehicle_data.get('vehicleCategory', 'unknown')}) validation failed: {str(e)}")
+                raise ValueError(
+                    f"Vehicle category {idx} ({vehicle_data.get('vehicleCategory', 'unknown')}) validation failed: {str(e)}")
             vehicle = serializer.save()
             created_vehicle_categories.append(vehicle)
             processed_data[f'vehicleCategoryData{idx}'] = vehicle.id
@@ -433,20 +523,25 @@ class AnalysisView(APIView):
         with transaction.atomic():
             try:
                 if created_load_categories:
-                    LoadCategoryModel.objects.filter(id__in=[cat.id for cat in created_load_categories]).delete()
-                    logger.info(f"Deleted {len(created_load_categories)} load categories")
+                    LoadCategoryModel.objects.filter(
+                        id__in=[cat.id for cat in created_load_categories]).delete()
+                    logger.info(
+                        f"Deleted {len(created_load_categories)} load categories")
             except Exception as e:
                 logger.error(f"Error deleting load categories: {e}")
             try:
                 if created_vehicle_categories:
-                    VehicleCategoryModel.objects.filter(id__in=[veh.id for veh in created_vehicle_categories]).delete()
-                    logger.info(f"Deleted {len(created_vehicle_categories)} vehicle categories")
+                    VehicleCategoryModel.objects.filter(
+                        id__in=[veh.id for veh in created_vehicle_categories]).delete()
+                    logger.info(
+                        f"Deleted {len(created_vehicle_categories)} vehicle categories")
             except Exception as e:
                 logger.error(f"Error deleting vehicle categories: {e}")
             try:
                 if ev_instance:
                     ev_instance.delete()
-                    logger.info(f"Deleted EV analysis instance {ev_instance.id}")
+                    logger.info(
+                        f"Deleted EV analysis instance {ev_instance.id}")
             except Exception as e:
                 logger.error(f"Error deleting EV analysis instance: {e}")
             try:
@@ -492,7 +587,7 @@ class AnalysisView(APIView):
         analysis_data = {
             'resolution': ev_instance.resolution,
             'BR_F': ev_instance.BR_F,
-            'sharedSavaing': ev_instance.shared_saving,  # Corrected typo if model uses shared_saving
+            'sharedSavaing': ev_instance.shared_saving,
             'sum_pk_cost': ev_instance.sum_pk_cost,
             'sum_zero_cost': ev_instance.sum_zero_cost,
             'sum_op_cost': ev_instance.sum_op_cost,
@@ -549,7 +644,8 @@ class AnalysisView(APIView):
             if e <= 0:
                 raise ValueError("Efficiency (e) must be positive")
             Total_Charges = n * f
-            Blocks = np.arange(1, int(1440/input_data['resolution'])+1, 1).reshape((1, int(1440/input_data['resolution'])))
+            Blocks = np.arange(1, int(
+                1440/input_data['resolution'])+1, 1).reshape((1, int(1440/input_data['resolution'])))
             ex1 = np.arange(Blocks.min(), Blocks.max()+1, 1)
             mu = math.ceil(k/input_data['resolution'])
             sigma = max(math.ceil(l/input_data['resolution']), 1)
@@ -568,19 +664,26 @@ class AnalysisView(APIView):
             EndingSOC_Prob = ss.norm.pdf(EndingSOC, mu3, sigma3)
             dummy = np.tile(StartingSOC, (101, 1))
             dummy_transpose = dummy.transpose()
-            StartingSOC_Matrix = np.tile(dummy_transpose, (int(1440/input_data['resolution']), 1))
-            EndingSOC_Matrix = np.tile(EndingSOC, ((int(1440/input_data['resolution']))*(r+1), 1))
-            EndingSOC_Prob_Matrix = np.tile(EndingSOC_Prob, ((int(1440/input_data['resolution']))*(r+1), 1))
-            ATD_Column = np.reshape(ATD, ((int(1440/input_data['resolution']))*(r+1), 1))
+            StartingSOC_Matrix = np.tile(
+                dummy_transpose, (int(1440/input_data['resolution']), 1))
+            EndingSOC_Matrix = np.tile(
+                EndingSOC, ((int(1440/input_data['resolution']))*(r+1), 1))
+            EndingSOC_Prob_Matrix = np.tile(
+                EndingSOC_Prob, ((int(1440/input_data['resolution']))*(r+1), 1))
+            ATD_Column = np.reshape(
+                ATD, ((int(1440/input_data['resolution']))*(r+1), 1))
             Veh_All_Comb = ATD_Column * EndingSOC_Prob_Matrix
-            Charging_Duration = ((60*c/input_data['resolution'])/(p*e/100)) * (EndingSOC_Matrix-StartingSOC_Matrix)
-            Charging_Duration_P = np.where(Charging_Duration < 0, 0, Charging_Duration)
+            Charging_Duration = (
+                (60*c/input_data['resolution'])/(p*e/100)) * (EndingSOC_Matrix-StartingSOC_Matrix)
+            Charging_Duration_P = np.where(
+                Charging_Duration < 0, 0, Charging_Duration)
             Output = np.sum(Veh_All_Comb, axis=1)
             Blo_sum_linear = np.zeros(int(1440/input_data['resolution']))
             for i, value in enumerate(Output):
                 block_idx = i % int(1440/input_data['resolution'])
                 Blo_sum_linear[block_idx] += value
-            Blo_load_sec = (p * Blo_sum_linear).reshape(1, int(1440/input_data['resolution']))
+            Blo_load_sec = (p * Blo_sum_linear).reshape(1,
+                                                        int(1440/input_data['resolution']))
             return Blo_load_sec.tolist()
 
         dracula = []
@@ -589,7 +692,8 @@ class AnalysisView(APIView):
                 res = load_forecast(**vehicle)
                 dracula.append(res)
             except Exception as e:
-                logger.error(f"Load forecast failed for vehicle {vehicle.get('vehicleCategory', 'unknown')}: {str(e)}")
+                logger.error(
+                    f"Load forecast failed for vehicle {vehicle.get('vehicleCategory', 'unknown')}: {str(e)}")
                 raise
         ddf = pd.DataFrame(np.concatenate(dracula))
         new_dict = {
@@ -601,7 +705,8 @@ class AnalysisView(APIView):
             new_dict[1][cat] = category['specifySplit']
             new_dict[2][cat] = category['salesCAGR']
         try:
-            excelimport = pd.read_excel(input_data['isLoadSplitFile'], header=None)
+            excelimport = pd.read_excel(
+                input_data['isLoadSplitFile'], header=None)
             if excelimport.shape[0] < 5:
                 raise ValueError("Excel file must have at least 5 rows")
             if excelimport.shape[1] != 11:
@@ -609,11 +714,14 @@ class AnalysisView(APIView):
         except Exception as e:
             raise ValueError(f"Failed to read Excel file: {str(e)}")
         Source = excelimport.iloc[4:, :].copy()
-        Source.columns = ['Meter.No', 'datetime_utc', 'Active_B_PH', 'Active_Y_PH', 'Active_R_PH', 'Reactive_B_PH', 'Reactive_Y_PH', 'Reactive_R_PH', 'VBV', 'VYV', 'VRV']
+        Source.columns = ['Meter.No', 'datetime_utc', 'Active_B_PH', 'Active_Y_PH',
+                          'Active_R_PH', 'Reactive_B_PH', 'Reactive_Y_PH', 'Reactive_R_PH', 'VBV', 'VYV', 'VRV']
         Source = Source.reset_index(drop=True)
         try:
-            Source[''] = ((Source.Active_B_PH * Source.VBV) + (Source.Active_Y_PH * Source.VYV) + (Source.Active_R_PH * Source.VRV))/1000
-            Source['datetime_utc'] = pd.to_datetime(Source['datetime_utc'], errors='coerce')
+            Source[''] = ((Source.Active_B_PH * Source.VBV) + (Source.Active_Y_PH *
+                          Source.VYV) + (Source.Active_R_PH * Source.VRV))/1000
+            Source['datetime_utc'] = pd.to_datetime(
+                Source['datetime_utc'], errors='coerce')
             if Source['datetime_utc'].isna().any():
                 raise ValueError("Invalid datetime values in Excel file")
             Source['date'] = Source['datetime_utc'].dt.date
@@ -622,13 +730,17 @@ class AnalysisView(APIView):
         labels = Source['datetime_utc'].dt.date.unique()
         slots = pd.DataFrame(Source['datetime_utc'].dt.time.unique())
         slots.columns = ['slot_labels']
-        value_to_repeat = excelimport.iloc[0, 1] * (float(input_data['BR_F'])/100)
+        value_to_repeat = excelimport.iloc[0,
+                                           1] * (float(input_data['BR_F'])/100)
         number_of_repeats = int(1440/input_data['resolution'])
-        Transformer_capacity = pd.DataFrame(np.repeat(value_to_repeat, number_of_repeats))
+        Transformer_capacity = pd.DataFrame(
+            np.repeat(value_to_repeat, number_of_repeats))
         Transformer_capacity.columns = ['Transformer safety planning trigger']
-        Transformer_capacity['100% Transformer rated capacity'] = np.repeat(excelimport.iloc[0, 1], number_of_repeats)
+        Transformer_capacity['100% Transformer rated capacity'] = np.repeat(
+            excelimport.iloc[0, 1], number_of_repeats)
         Source.set_index('datetime_utc', inplace=True)
-        Calculated_load = Source.drop(['Meter.No', 'Active_B_PH', 'Active_Y_PH', 'Active_R_PH', 'Reactive_B_PH', 'Reactive_Y_PH', 'Reactive_R_PH', 'VBV', 'VYV', 'VRV'], axis=1)
+        Calculated_load = Source.drop(['Meter.No', 'Active_B_PH', 'Active_Y_PH', 'Active_R_PH',
+                                      'Reactive_B_PH', 'Reactive_Y_PH', 'Reactive_R_PH', 'VBV', 'VYV', 'VRV'], axis=1)
         Calculated_load.index.name = None
         abc = pd.DataFrame(Calculated_load[''].astype(float).values)
         resolution = input_data['resolution']
@@ -637,11 +749,13 @@ class AnalysisView(APIView):
         complete_days = total_data_points // time_blocks_per_day
         if total_data_points % time_blocks_per_day != 0:
             abc_trimmed = abc.iloc[:complete_days * time_blocks_per_day]
-            logger.warning(f"Data trimmed from {total_data_points} to {len(abc_trimmed)} points to fit {complete_days} complete days")
+            logger.warning(
+                f"Data trimmed from {total_data_points} to {len(abc_trimmed)} points to fit {complete_days} complete days")
         else:
             abc_trimmed = abc
         try:
-            load_extract = pd.DataFrame(np.reshape(abc_trimmed.to_numpy(), (complete_days, time_blocks_per_day)))
+            load_extract = pd.DataFrame(np.reshape(
+                abc_trimmed.to_numpy(), (complete_days, time_blocks_per_day)))
             load_extract = load_extract.T
         except ValueError as e:
             logger.error(f"Reshape error: {e}")
@@ -649,24 +763,32 @@ class AnalysisView(APIView):
         if len(labels) != load_extract.shape[1]:
             if len(labels) > load_extract.shape[1]:
                 labels = labels[:load_extract.shape[1]]
-                logger.warning(f"Trimmed labels to {len(labels)} to match data columns")
+                logger.warning(
+                    f"Trimmed labels to {len(labels)} to match data columns")
             else:
                 additional_needed = load_extract.shape[1] - len(labels)
-                labels_to_repeat = labels[-additional_needed:] if additional_needed <= len(labels) else labels
-                labels = np.concatenate([labels, labels_to_repeat[:additional_needed]])
-                logger.warning(f"Extended labels to {len(labels)} to match data columns")
+                labels_to_repeat = labels[-additional_needed:] if additional_needed <= len(
+                    labels) else labels
+                labels = np.concatenate(
+                    [labels, labels_to_repeat[:additional_needed]])
+                logger.warning(
+                    f"Extended labels to {len(labels)} to match data columns")
         final_load = load_extract.copy()
         final_load.columns = labels
         max_cols = min(406, final_load.shape[1])
         selected_range = final_load.iloc[:, :max_cols].copy()
         selected_ranges = [selected_range]
         for year in range(1, 5):
-            growth_factors = {cat: new_dict[1][cat] * (1 + new_dict[2][cat] / 100) for cat in new_dict[1]}
-            next_range = (selected_ranges[-1]/100) * sum(growth_factors.values())
+            growth_factors = {
+                cat: new_dict[1][cat] * (1 + new_dict[2][cat] / 100) for cat in new_dict[1]}
+            next_range = (selected_ranges[-1]/100) * \
+                sum(growth_factors.values())
             selected_ranges.append(next_range)
-        selected_range_2, selected_range_3, selected_range_4, selected_range_5 = selected_ranges[1:5]
+        selected_range_2, selected_range_3, selected_range_4, selected_range_5 = selected_ranges[
+            1:5]
         shazam = pd.DataFrame(ddf.sum(axis=0))
-        growth_factors = [(input_data['vehicleCategoryData'][i]['CAGR_V'])/100 + 1 for i in range(len(input_data['vehicleCategoryData']))]
+        growth_factors = [(input_data['vehicleCategoryData'][i]['CAGR_V']) /
+                          100 + 1 for i in range(len(input_data['vehicleCategoryData']))]
         ddf2 = ddf.mul(growth_factors, axis=0)
         shazam_2 = pd.DataFrame(ddf2.sum(axis=0).to_numpy())
         ddf3 = ddf2.mul(growth_factors, axis=0)
@@ -710,7 +832,8 @@ class AnalysisView(APIView):
             ddf_years.append(ddf_next)
             shazam = pd.DataFrame(ddf_next.sum(axis=0).to_numpy())
             shazam_years.append(shazam)
-            ev_loads.append({f'Year {year}': ddf_next.sum(axis=0).to_numpy().flatten().tolist()})
+            ev_loads.append({f'Year {year}': ddf_next.sum(
+                axis=0).to_numpy().flatten().tolist()})
         output_data['EV_Load'] = ev_loads
         try:
             start_date = "2019-05-01"
@@ -723,7 +846,8 @@ class AnalysisView(APIView):
         pivot_df = pd.DataFrame(
             index=time_slots, columns=dates.date, dtype=float)
         for i, date in enumerate(dates):
-            base_pattern = np.array([60 + 40 * np.sin(2 * np.pi * (h) / 24) + np.random.normal(0, 5) for h in np.linspace(0, 24, 48)])
+            base_pattern = np.array(
+                [60 + 40 * np.sin(2 * np.pi * (h) / 24) + np.random.normal(0, 5) for h in np.linspace(0, 24, 48)])
             if date.weekday() < 5:
                 base_pattern *= 1.0 + 0.1 * np.random.random()
             else:
@@ -738,7 +862,8 @@ class AnalysisView(APIView):
             daily_load += np.random.normal(0, 3, size=len(daily_load))
             pivot_df[date.date()] = np.round(daily_load.astype(float), 6)
         output_data['DT_Base_Load'] = pivot_df.values.tolist()
-        selected_ranges = [selected_range, selected_range_2, selected_range_3, selected_range_4, selected_range_5]
+        selected_ranges = [selected_range, selected_range_2,
+                           selected_range_3, selected_range_4, selected_range_5]
         base_loads = []
         for year in range(5):
             if year >= len(selected_ranges):
@@ -749,17 +874,24 @@ class AnalysisView(APIView):
             mean_load = base_load.mean(axis=1).values.tolist()
             base_loads.append({f'Year {year+1}': mean_load})
         output_data['Base_Load'] = base_loads
-        required_net_loads = [pd.DataFrame(selected_ranges[year]) + shazam_years[year].values for year in range(5)]
+        required_net_loads = [pd.DataFrame(
+            selected_ranges[year]) + shazam_years[year].values for year in range(5)]
         combined_loads = []
         for year in range(5):
-            base_load = pd.DataFrame(selected_ranges[year]).mean(axis=1).values.tolist()
+            base_load = pd.DataFrame(selected_ranges[year]).mean(
+                axis=1).values.tolist()
             ev_load = shazam_years[year].mean(axis=1).values.tolist()
-            combined_loads.append({f'Year {year+1}': {'base_load': base_load, 'ev_load': ev_load}})
+            combined_loads.append(
+                {f'Year {year+1}': {'base_load': base_load, 'ev_load': ev_load}})
         output_data['Base_EV_Load'] = combined_loads
-        final_res_res = pd.DataFrame(np.random.rand(5, int(1440/input_data['resolution'])))
-        output_data['Base_ToD_EV_Load'] = self.generate_tod_ev_load_plot(final_res_res, excelimport, input_data['resolution'])
-        overshots = [rn - (excelimport.iloc[0, 1] * (float(input_data['BR_F'])/100)) for rn in required_net_loads]
-        overshots_r = [rn - (excelimport.iloc[0, 1] * (90/100)) for rn in required_net_loads]
+        final_res_res = pd.DataFrame(np.random.rand(
+            5, int(1440/input_data['resolution'])))
+        output_data['Base_ToD_EV_Load'] = self.generate_tod_ev_load_plot(
+            final_res_res, excelimport, input_data['resolution'])
+        overshots = [rn - (excelimport.iloc[0, 1] * (float(input_data['BR_F'])/100))
+                     for rn in required_net_loads]
+        overshots_r = [rn - (excelimport.iloc[0, 1] * (90/100))
+                       for rn in required_net_loads]
         table_data = []
         for year in range(5):
             table_data.append({
@@ -784,9 +916,11 @@ class AnalysisView(APIView):
                 'variable': 'dates',
                 'value': f'overshot_rated_year_{year}'
             })
-            overshot_data.append({f'Year {year}': {'planning': ov_p.to_dict('records'), 'rated': ov_r_p.to_dict('records')}})
+            overshot_data.append({f'Year {year}': {'planning': ov_p.to_dict(
+                'records'), 'rated': ov_r_p.to_dict('records')}})
         output_data['Overshot'] = overshot_data
-        abc_ddf = pd.DataFrame({'seasons': np.random.choice(['summer', 'winter'], 5), 'unit_type': np.random.choice(['pk', 'op', '0'], 5)})
+        abc_ddf = pd.DataFrame({'seasons': np.random.choice(
+            ['summer', 'winter'], 5), 'unit_type': np.random.choice(['pk', 'op', '0'], 5)})
         conditions_4 = [
             (abc_ddf['seasons'] == 'summer') & (abc_ddf['unit_type'] == 'pk'),
             (abc_ddf['seasons'] == 'winter') & (abc_ddf['unit_type'] == 'pk'),
@@ -803,45 +937,66 @@ class AnalysisView(APIView):
         abc_ddf['utility_proc_tariff'] = np.select(conditions_4, values_4)
         ddf_x = final_res_res.copy()
         ddf_x.index = abc_ddf.index
-        old_utility_cost = (ddf_x.mul(abc_ddf['utility_proc_tariff'], axis=0) / 2)
+        old_utility_cost = (
+            ddf_x.mul(abc_ddf['utility_proc_tariff'], axis=0) / 2)
         old_utility_cost.columns = ddf_x.columns.astype(str)
-        old_utility_cost = old_utility_cost.rename(columns=lambda x: x + '_old_cost')
-        final_res = pd.DataFrame(np.random.rand(5, int(1440 / input_data['resolution'])))
-        new_utility_cost = (final_res.mul(abc_ddf['utility_proc_tariff'], axis=0) / 2)
+        old_utility_cost = old_utility_cost.rename(
+            columns=lambda x: x + '_old_cost')
+        final_res = pd.DataFrame(np.random.rand(
+            5, int(1440 / input_data['resolution'])))
+        new_utility_cost = (final_res.mul(
+            abc_ddf['utility_proc_tariff'], axis=0) / 2)
         new_utility_cost.columns = final_res.columns.astype(str)
-        new_utility_cost = new_utility_cost.rename(columns=lambda x: x + '_new_cost')
-        retail_tariff_df = pd.DataFrame({'1': [1], '2': [1], '3': [1], '4': [1]})
+        new_utility_cost = new_utility_cost.rename(
+            columns=lambda x: x + '_new_cost')
+        retail_tariff_df = pd.DataFrame(
+            {'1': [1], '2': [1], '3': [1], '4': [1]})
         retail_tariff_value = retail_tariff_df.iloc[0].mean()
         old_tariff_revenue = (ddf_x * retail_tariff_value / 2)
         old_tariff_revenue.columns = ddf_x.columns.astype(str)
-        old_tariff_revenue = old_tariff_revenue.rename(columns=lambda x: x + '_old_tariff')
-        abc_ddf = pd.concat([abc_ddf, old_utility_cost, new_utility_cost, old_tariff_revenue], axis=1)
-        output_data['Load_Simulation_ToD_Calculation'] = abc_ddf.to_dict('records')
+        old_tariff_revenue = old_tariff_revenue.rename(
+            columns=lambda x: x + '_old_tariff')
+        abc_ddf = pd.concat(
+            [abc_ddf, old_utility_cost, new_utility_cost, old_tariff_revenue], axis=1)
+        output_data['Load_Simulation_ToD_Calculation'] = abc_ddf.to_dict(
+            'records')
 
         def calc_TOD_x(year_num):
             try:
-                s_pk_sum = abc_ddf.loc[(abc_ddf['seasons'] == 'summer') & (abc_ddf['unit_type'] == 'pk'), f'{year_num}_old_cost'].sum()
-                w_pk_sum = abc_ddf.loc[(abc_ddf['seasons'] == 'winter') & (abc_ddf['unit_type'] == 'pk'), f'{year_num}_old_cost'].sum()
-                s_op_sum = abc_ddf.loc[(abc_ddf['seasons'] == 'summer') & (abc_ddf['unit_type'] == 'op'), f'{year_num}_old_cost'].sum()
-                w_op_sum = abc_ddf.loc[(abc_ddf['seasons'] == 'winter') & (abc_ddf['unit_type'] == 'op'), f'{year_num}_old_cost'].sum()
-                s_0_sum = abc_ddf.loc[(abc_ddf['seasons'] == 'summer') & (abc_ddf['unit_type'] == '0'), f'{year_num}_old_cost'].sum()
-                w_0_sum = abc_ddf.loc[(abc_ddf['seasons'] == 'winter') & (abc_ddf['unit_type'] == '0'), f'{year_num}_old_cost'].sum()
+                s_pk_sum = abc_ddf.loc[(abc_ddf['seasons'] == 'summer') & (
+                    abc_ddf['unit_type'] == 'pk'), f'{year_num}_old_cost'].sum()
+                w_pk_sum = abc_ddf.loc[(abc_ddf['seasons'] == 'winter') & (
+                    abc_ddf['unit_type'] == 'pk'), f'{year_num}_old_cost'].sum()
+                s_op_sum = abc_ddf.loc[(abc_ddf['seasons'] == 'summer') & (
+                    abc_ddf['unit_type'] == 'op'), f'{year_num}_old_cost'].sum()
+                w_op_sum = abc_ddf.loc[(abc_ddf['seasons'] == 'winter') & (
+                    abc_ddf['unit_type'] == 'op'), f'{year_num}_old_cost'].sum()
+                s_0_sum = abc_ddf.loc[(abc_ddf['seasons'] == 'summer') & (
+                    abc_ddf['unit_type'] == '0'), f'{year_num}_old_cost'].sum()
+                w_0_sum = abc_ddf.loc[(abc_ddf['seasons'] == 'winter') & (
+                    abc_ddf['unit_type'] == '0'), f'{year_num}_old_cost'].sum()
                 old_tariff_sum = abc_ddf[f'{year_num}_old_tariff'].sum()
-                cost_diff = abc_ddf[f'{year_num}_new_cost'].sum() - abc_ddf[f'{year_num}_old_cost'].sum()
+                cost_diff = abc_ddf[f'{year_num}_new_cost'].sum(
+                ) - abc_ddf[f'{year_num}_old_cost'].sum()
                 shared_saving = float(input_data['sharedSavaing']) / 100
                 if retail_tariff_df.iloc[0, 0] == 0:
                     raise ValueError("Retail tariff cannot be zero")
-                numerator = (60 / input_data['resolution'] * (old_tariff_sum + cost_diff) * shared_saving) / retail_tariff_df.iloc[0, 0]
+                numerator = (60 / input_data['resolution'] * (
+                    old_tariff_sum + cost_diff) * shared_saving) / retail_tariff_df.iloc[0, 0]
                 denominator = (s_pk_sum + w_pk_sum - s_op_sum - w_op_sum)
                 if denominator == 0:
-                    logger.warning(f"Zero denominator in calc_TOD_x for year {year_num}, returning 0")
+                    logger.warning(
+                        f"Zero denominator in calc_TOD_x for year {year_num}, returning 0")
                     return 0
-                result = 100 * (numerator - (s_pk_sum + w_pk_sum + s_op_sum + w_op_sum + s_0_sum + w_0_sum)) / denominator
+                result = 100 * (numerator - (s_pk_sum + w_pk_sum +
+                                s_op_sum + w_op_sum + s_0_sum + w_0_sum)) / denominator
                 return round(result, 2)
             except Exception as e:
-                logger.error(f"Error in calc_TOD_x for year {year_num}: {str(e)}")
+                logger.error(
+                    f"Error in calc_TOD_x for year {year_num}: {str(e)}")
                 return 0
-        output_data['TOD_Surcharge_Rebate'] = [calc_TOD_x(i) for i in range(1, 5)]
+        output_data['TOD_Surcharge_Rebate'] = [
+            calc_TOD_x(i) for i in range(1, 5)]
         return output_data
 
     def generate_tod_ev_load_plot(self, final_res_res, excelimport, resolution):
@@ -951,4 +1106,4 @@ class PermanentAnalysisListCreateAPI(generics.ListCreateAPIView):
 class PermanentAnalysisRetrieveUpdateDestroyAPI(generics.RetrieveUpdateDestroyAPIView):
     queryset = PermanentAnalysis.objects.all()
     serializer_class = PermanentAnalysisSerializer
-    permission_classes = [IsAdminUser]
+    permission_classes = [permissions.AllowAny]
